Moved all LuceneServer handlers into handler package
