+++ src/main/java/com/yelp/nrtsearch/server/grpc/LuceneServer.java
@@ -29,45 +29,45 @@ import com.google.protobuf.util.JsonFormat;
 import com.yelp.nrtsearch.LuceneServerModule;
 import com.yelp.nrtsearch.server.config.LuceneServerConfiguration;
 import com.yelp.nrtsearch.server.config.QueryCacheConfig;
-import com.yelp.nrtsearch.server.luceneserver.AddDocumentHandler;
 import com.yelp.nrtsearch.server.luceneserver.AddReplicaHandler;
 import com.yelp.nrtsearch.server.luceneserver.CopyFilesHandler;
-import com.yelp.nrtsearch.server.luceneserver.CreateSnapshotHandler;
-import com.yelp.nrtsearch.server.luceneserver.DeleteAllDocumentsHandler;
-import com.yelp.nrtsearch.server.luceneserver.DeleteByQueryHandler;
-import com.yelp.nrtsearch.server.luceneserver.DeleteDocumentsHandler;
-import com.yelp.nrtsearch.server.luceneserver.DeleteIndexHandler;
 import com.yelp.nrtsearch.server.luceneserver.GetNodesInfoHandler;
-import com.yelp.nrtsearch.server.luceneserver.GetStateHandler;
 import com.yelp.nrtsearch.server.luceneserver.GlobalState;
 import com.yelp.nrtsearch.server.luceneserver.IndexState;
-import com.yelp.nrtsearch.server.luceneserver.LiveSettingsHandler;
 import com.yelp.nrtsearch.server.luceneserver.NewNRTPointHandler;
 import com.yelp.nrtsearch.server.luceneserver.RecvCopyStateHandler;
-import com.yelp.nrtsearch.server.luceneserver.ReleaseSnapshotHandler;
 import com.yelp.nrtsearch.server.luceneserver.ReplicaCurrentSearchingVersionHandler;
-import com.yelp.nrtsearch.server.luceneserver.SearchHandler;
-import com.yelp.nrtsearch.server.luceneserver.SettingsHandler;
 import com.yelp.nrtsearch.server.luceneserver.ShardState;
 import com.yelp.nrtsearch.server.luceneserver.WriteNRTPointHandler;
 import com.yelp.nrtsearch.server.luceneserver.analysis.AnalyzerCreator;
 import com.yelp.nrtsearch.server.luceneserver.custom.request.CustomRequestProcessor;
 import com.yelp.nrtsearch.server.luceneserver.field.FieldDefCreator;
+import com.yelp.nrtsearch.server.luceneserver.handler.AddDocumentHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.BackupWarmingQueriesHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.CommitHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.CreateIndexHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.CreateSnapshotHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.CustomHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.DeleteAllDocumentsHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.DeleteByQueryHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.DeleteDocumentsHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.DeleteIndexHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.ForceMergeDeletesHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.ForceMergeHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.GetAllSnapshotIndexGenHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.GetStateHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.GlobalStateHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.IndicesHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.LiveSettingsHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.MetricsHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.ReadyHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.RefreshHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.RegisterFieldsHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.ReleaseSnapshotHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.ReloadStateHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.SearchHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.SearchV2Handler;
+import com.yelp.nrtsearch.server.luceneserver.handler.SettingsHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StartIndexHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StartIndexV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StatsHandler;
