--- src/main/java/com/yelp/nrtsearch/server/luceneserver/StopIndexHandler.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2020 Yelp Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.yelp.nrtsearch.server.luceneserver;
-
-import com.yelp.nrtsearch.server.grpc.DummyResponse;
-import com.yelp.nrtsearch.server.grpc.StopIndexRequest;
-import java.io.IOException;
-
-public class StopIndexHandler implements Handler<StopIndexRequest, DummyResponse> {
-  @Override
-  public DummyResponse handle(IndexState indexState, StopIndexRequest stopIndexRequest)
-      throws HandlerException {
-    try {
-      indexState.close();
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    return DummyResponse.newBuilder().setOk("ok").build();
-  }
-}
