+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/field/ObjectFieldDef.java
@@ -24,9 +24,9 @@ import com.google.protobuf.Value;
 import com.google.protobuf.util.JsonFormat;
 import com.yelp.nrtsearch.server.grpc.Field;
 import com.yelp.nrtsearch.server.grpc.SearchResponse;
-import com.yelp.nrtsearch.server.luceneserver.AddDocumentHandler;
 import com.yelp.nrtsearch.server.luceneserver.IndexState;
 import com.yelp.nrtsearch.server.luceneserver.doc.LoadedDocValues;
+import com.yelp.nrtsearch.server.luceneserver.handler.AddDocumentHandler;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
