+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/handler/GetStateHandler.java
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yelp.nrtsearch.server.luceneserver;
+package com.yelp.nrtsearch.server.luceneserver.handler;
 
 import com.google.gson.JsonObject;
 import com.yelp.nrtsearch.server.grpc.StateRequest;
 import com.yelp.nrtsearch.server.grpc.StateResponse;
-import com.yelp.nrtsearch.server.luceneserver.handler.Handler;
+import com.yelp.nrtsearch.server.luceneserver.GlobalState;
+import com.yelp.nrtsearch.server.luceneserver.IndexState;
 import io.grpc.Status;
 import io.grpc.stub.StreamObserver;
 import java.io.IOException;
