+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/handler/SearchV2Handler.java
@@ -21,7 +21,6 @@ import com.google.protobuf.util.JsonFormat.Printer;
 import com.yelp.nrtsearch.server.grpc.SearchRequest;
 import com.yelp.nrtsearch.server.grpc.SearchResponse;
 import com.yelp.nrtsearch.server.luceneserver.GlobalState;
-import com.yelp.nrtsearch.server.luceneserver.SearchHandler;
 import com.yelp.nrtsearch.server.utils.ProtoMessagePrinter;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
