+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/search/sort/SortContext.java
@@ -18,9 +18,9 @@ package com.yelp.nrtsearch.server.luceneserver.search.sort;
 import com.yelp.nrtsearch.server.grpc.QuerySortField;
 import com.yelp.nrtsearch.server.grpc.SearchResponse.Hit.CompositeFieldValue;
 import com.yelp.nrtsearch.server.grpc.SortType;
-import com.yelp.nrtsearch.server.luceneserver.SearchHandler;
 import com.yelp.nrtsearch.server.luceneserver.field.FieldDef;
 import com.yelp.nrtsearch.server.luceneserver.field.properties.Sortable;
+import com.yelp.nrtsearch.server.luceneserver.handler.SearchHandler;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
